+++ src/main/java/net/mohron/skyclaims/command/team/CommandPromote.java
@@ -21,6 +21,7 @@ package net.mohron.skyclaims.command.team;
 import net.mohron.skyclaims.command.CommandBase;
 import net.mohron.skyclaims.command.CommandIsland;
 import net.mohron.skyclaims.permissions.Permissions;
+import net.mohron.skyclaims.team.Invite;
 import net.mohron.skyclaims.team.PrivilegeType;
 import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
@@ -35,26 +36,26 @@ import org.spongepowered.api.text.format.TextColors;
 
 import javax.annotation.Nonnull;
 
-public class CommandKick extends CommandBase.IslandCommand {
+public class CommandPromote extends CommandBase.IslandCommand {
 
-    public static final String HELP_TEXT = "used to remove players from an island.";
+    public static final String HELP_TEXT = "used to promote a player on an island.";
     private static final Text USER = Text.of("user");
 
     public static CommandSpec commandSpec = CommandSpec.builder()
-        .permission(Permissions.COMMAND_KICK)
+        .permission(Permissions.COMMAND_PROMOTE)
         .arguments(GenericArguments.user(USER))
         .description(Text.of(HELP_TEXT))
-        .executor(new CommandKick())
+        .executor(new CommandPromote())
         .build();
 
     public static void register() {
         try {
-            CommandIsland.addSubCommand(commandSpec, "kick");
+            CommandIsland.addSubCommand(commandSpec, "promote");
             PLUGIN.getGame().getCommandManager().register(PLUGIN, commandSpec);
-            PLUGIN.getLogger().debug("Registered command: CommandKick");
+            PLUGIN.getLogger().debug("Registered command: CommandPromote");
         } catch (UnsupportedOperationException e) {
             e.printStackTrace();
-            PLUGIN.getLogger().error("Failed to register command: CommandKick");
+            PLUGIN.getLogger().error("Failed to register command: CommandPromote");
         }
     }
 
@@ -64,26 +65,34 @@ public class CommandKick extends CommandBase.IslandCommand {
         if (user == null) {
             throw new CommandException(Text.of(TextColors.RED, "A user argument must be provided."));
         } else if (player.equals(user)) {
-            throw new CommandException(Text.of(TextColors.RED, "You cannot kick yourself!"));
-        } else if (island.getPrivilegeType(user) == PrivilegeType.NONE) {
-            throw new CommandException(Text.of(
-                PrivilegeType.NONE.format(user.getName()), TextColors.RED, " is not a member of ", island.getName(), TextColors.RED, "!"
-            ));
-        } else if (island.getPrivilegeType(player).ordinal() >= island.getPrivilegeType(user).ordinal()) {
-            throw new CommandException(Text.of(
-                TextColors.RED, "You do not have permission to kick ", island.getPrivilegeType(user).format(user.getName()),
-                " from ", island.getName(), TextColors.RED, "."
-            ));
-        }
-
-        user.getPlayer().ifPresent(p -> {
-            if (island.getPlayers().contains(p)) {
-                p.setLocationSafely(island.getWorld().getSpawnLocation());
+            throw new CommandException(Text.of(TextColors.RED, "You cannot promote yourself!"));
+        } else if (!island.isOwner(player)) {
+            throw new CommandException(Text.of(TextColors.RED, "You do not have permission to promote players on this island!"));
+        } else {
+            PrivilegeType type = island.getPrivilegeType(user);
+            if (type == PrivilegeType.MANAGER) {
+                Invite invite = Invite.builder()
+                    .island(island)
+                    .sender(player)
+                    .receiver(user)
+                    .privilegeType(type)
+                    .build();
+                if (PLUGIN.getInviteService().inviteExists(invite)) {
+                    throw new CommandException(Text.of(TextColors.RED, "Invite already exists!"));
+                } else {
+                    invite.send();
+                    player.sendMessage(Text.of(
+                        TextColors.GREEN, "Island ownership transfer request sent to ", type.format(user.getName()), TextColors.GREEN, "."
+                    ));
+                }
+            } else {
+                player.sendMessage(Text.of(
+                    type.format(user.getName()), TextColors.GREEN, " has been promoted from a ",
+                    island.getPrivilegeType(user).toText(), TextColors.GREEN, " to a ", type.toText(), TextColors.GREEN, "."
+                ));
+                island.promote(user);
             }
-            p.sendMessage(Text.of(TextColors.RED, "You have been removed from ", island.getName(), TextColors.RED, "!"));
-        });
-
-        island.removeMember(user);
+        }
 
         return CommandResult.success();
     }